src/websocket_handler.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

"""
WebSocket Handler Module
------------------------
Handles the WebSocket connection to the OKX L2 order book feed.
Processes incoming data and maintains the order book state.

"""

import json
import threading
import time
import logging
import websocket
from collections import deque

class WebSocketHandler:
    """Handles WebSocket connection and data processing for L2 order book."""
    
    def __init__(self, url):
        """
        Initialize the WebSocket handler.
        
        Args:
            url (str): WebSocket endpoint URL
        """
        self.url = url
        self.ws = None
        self.connected = False
        self.order_book = {"bids": {}, "asks": {}}
        self.data_buffer = deque(maxlen=1000)  # Store recent updates
        self.lock = threading.Lock()
        self.logger = logging.getLogger(__name__)
        
    def connect(self):
        """Establish WebSocket connection."""
        self.logger.info(f"Connecting to WebSocket: {self.url}")
        # Implementation will be added in Phase 2
        pass
        
    def on_message(self, ws, message):
        """Handle incoming WebSocket messages."""
        # Implementation will be added in Phase 2
        pass
        
    def on_error(self, ws, error):
        """Handle WebSocket errors."""
        self.logger.error(f"WebSocket error: {str(error)}")
        
    def on_close(self, ws, close_status_code, close_msg):
        """Handle WebSocket connection close."""
        self.logger.info("WebSocket connection closed")
        self.connected = False
        
    def on_open(self, ws):
        """Handle WebSocket connection open."""
        self.logger.info("WebSocket connection established")
        self.connected = True
        
    def get_current_order_book(self):
        """
        Get the current state of the order book.
        
        Returns:
            dict: Current order book with bids and asks
        """
        with self.lock:
            return self.order_book.copy()
        
    def close(self):
        """Close the WebSocket connection."""
        if self.ws and self.connected:
            self.ws.close()
            self.logger.info("WebSocket connection closed")