requirements.txt
# WebSocket and networking
websocket-client==1.6.1

# UI
# Note: tkinter comes with Python installation

# Data and computation
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
statsmodels==0.14.0

# Visualization
matplotlib==3.7.2

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3